DirectWrite font db: Fix fallback fonts
The implementation of this was based on a misunderstanding of the API. The
GetMatchingFonts() function does not give a sorted list of fonts that
can be used in place of the given family, but just requests the fonts in
the family itself.

Instead, we use the same implementation as in the other two font databases
we have on Windows (moving the implementation to be shared).

Change-Id: I6a7b73e3d8376f7d97f598db0d7b63122ad1940c
Reviewed-by: Lars Knoll <lars.knoll@qt.io>