a11y: Remove a ### Qt6 that doesn't make much sense
Since the enum names are "the same", the comment seems to suggest to
change the enum value of Canvas to 0x401. I don't see any benefit to
that really.

In addition, we are today using UIA in favor of MSAA, and the closest
matching role (controlTypeId) seems to be for a static image
(UIA_ImageControlTypeId), which is quite different from the semantic of
IA2_ROLE_CANVAS.

For the record, here's a list of similar roles for different a11y APIs:

IA2:
  IA2_ROLE_CANVAS        An object that can be drawn into and to manage
			 events from the objects drawn into it. Also
			 refer to IA2_ROLE_FRAME, IA2_ROLE_GLASS_PANE,
			 and IA2_ROLE_LAYERED_PANE.

MSAA:
  ROLE_SYSTEM_DIAGRAM    The object represents a graphical image that is
                         used to diagram data.

UIA:
  UIA_ImageControlTypeId Identifies the _Image_ control type.

atspi:
  ATSPI_ROLE_CANVAS      Object that can be drawn into and is used to
                         trap events.

Change-Id: Ic2ead0dc40be0ae2e798f49285eb6a392cc29142
Reviewed-by: André de la Rocha <andre.rocha@qt.io>