Improve error message when mixing incompatible Qt library versions
The implementation of the check has been moved to a single helper
function. The reason for doing this check in QWidgetPrivate as well,
when it's already done in QObjectPrivate, is to catch incompatible
libraries where QtWidgets is the odd one out, e.g.:

 QtSomeModule 5.15.0 -> QtWidget 5.15.1 -> QtCore 5.15.0

Technically any non-final subclass of QObjectPrivate should have
this check.

Change-Id: Ia74064ad27de7335040a6d6b37d11574f818c878
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>